+++ core/js-native/src/main/scala/cats/effect/unsafe/metrics/IORuntimeMetricsPlatform.scala
@@ -14,10 +14,6 @@
  * limitations under the License.
  */
 
-package cats.effect.metrics
+package cats.effect.unsafe.metrics
 
-private[metrics] trait IORuntimeMetricsPlatform { this: IORuntimeMetrics =>
-
-  def workStealingThreadPool: Option[WorkStealingPoolMetrics]
-
-}
+private[metrics] trait IORuntimeMetricsPlatform { this: IORuntimeMetrics => }
