+++ core/jvm/src/main/scala/cats/effect/unsafe/WorkStealingThreadPool.scala
@@ -81,7 +81,7 @@ private[effect] final class WorkStealingThreadPool[P <: AnyRef](
    * References to worker threads and their local queues.
    */
   private[this] val workerThreads: Array[WorkerThread[P]] = new Array(threadCount)
-  private[effect] val localQueues: Array[LocalQueue] = new Array(threadCount)
+  private[unsafe] val localQueues: Array[LocalQueue] = new Array(threadCount)
   private[unsafe] val sleepers: Array[TimerHeap] = new Array(threadCount)
   private[unsafe] val parkedSignals: Array[AtomicBoolean] = new Array(threadCount)
   private[unsafe] val fiberBags: Array[WeakBag[Runnable]] = new Array(threadCount)
@@ -778,7 +778,7 @@ private[effect] final class WorkStealingThreadPool[P <: AnyRef](
    * @return
    *   the number of worker threads backing the compute pool
    */
-  private[effect] def getWorkerThreadCount(): Int =
+  private[unsafe] def getWorkerThreadCount(): Int =
     threadCount
 
   /**
@@ -788,7 +788,7 @@ private[effect] final class WorkStealingThreadPool[P <: AnyRef](
    * @return
    *   the number of active worker threads
    */
-  private[effect] def getActiveThreadCount(): Int = {
+  private[unsafe] def getActiveThreadCount(): Int = {
     val st = state.get()
     (st & UnparkMask) >>> UnparkShift
   }
@@ -800,7 +800,7 @@ private[effect] final class WorkStealingThreadPool[P <: AnyRef](
    * @return
    *   the number of worker threads searching for work
    */
-  private[effect] def getSearchingThreadCount(): Int = {
+  private[unsafe] def getSearchingThreadCount(): Int = {
     val st = state.get()
     st & SearchMask
   }
@@ -812,7 +812,7 @@ private[effect] final class WorkStealingThreadPool[P <: AnyRef](
    * @return
    *   the number of blocked worker threads
    */
-  private[effect] def getBlockedWorkerThreadCount(): Int =
+  private[unsafe] def getBlockedWorkerThreadCount(): Int =
     blockedWorkerThreadCounter.get()
 
   /**
@@ -821,7 +821,7 @@ private[effect] final class WorkStealingThreadPool[P <: AnyRef](
    * @return
    *   the total number of fibers enqueued on all local queues
    */
-  private[effect] def getLocalQueueFiberCount(): Long =
+  private[unsafe] def getLocalQueueFiberCount(): Long =
     localQueues.map(_.size().toLong).sum
 
   /**
@@ -834,7 +834,7 @@ private[effect] final class WorkStealingThreadPool[P <: AnyRef](
    * @return
    *   the number of asynchronously suspended fibers
    */
-  private[effect] def getSuspendedFiberCount(): Long =
+  private[unsafe] def getSuspendedFiberCount(): Long =
     workerThreads.map(_.getSuspendedFiberCount().toLong).sum
 }
